+++ test/aqua.jl
@@ -18,7 +18,6 @@ using Aqua
     # If the number of ambiguities is less than the limit below,
     # then please lower the limit based on the new number of ambiguities.
     # We're trying to drive this number down to zero to reduce latency.
-    # Uncomment for debugging:
     for method_ambiguity in ambs
         @show method_ambiguity
     end